refactor(stripe): integrate Stripe service methods into StripeService class and update user credits handling
