--- web-app/src/lib/services/stripe/index.ts
@@ -1,1 +1,0 @@
-export * from "./service";
